requirements.txt
# === LangChain & LangGraph (v0.3 라인 호환 세트) ===
langchain==0.3.27
langchain-core==0.3.79
langchain-community==0.3.31
langchain-text-splitters==0.3.11
langgraph==0.2.25
langchain-huggingface==0.1.1

# === Transformers & HuggingFace ===
transformers>=4.48,<4.50
huggingface_hub==0.25.2
accelerate>=0.26.0
sentence-transformers>=3.0.0
datasets>=2.21.0
# tokenizers는 transitive dep이므로 별도 지정 X

# === Retriever / Search ===
faiss-cpu==1.8.0
rank-bm25==0.2.2
tavily-python==0.4.0

# === 문서 처리 ===
pymupdf==1.24.4
python-docx>=1.1.2

# === 일반 유틸 ===
pydantic>=2.6.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0
regex>=2023.6.3
streamlit>=1.38,<1.41

openai==1.40.6
